+++ engine/apps/alerts/tests/test_paging.py
@@ -4,8 +4,14 @@ import pytest
 from django.utils import timezone
 
 from apps.alerts.models import AlertGroup, AlertGroupLogRecord, UserHasNotification
-from apps.alerts.paging import PagingError, check_user_availability, direct_paging, unpage_user
-from apps.base.models import UserNotificationPolicy
+from apps.alerts.paging import (
+    DirectPagingUserTeamValidationError,
+    _construct_title,
+    direct_paging,
+    integration_is_notifiable,
+    unpage_user,
+    user_is_oncall,
+)
 from apps.schedules.models import CustomOnCallShift, OnCallScheduleWeb
 
 
@@ -16,12 +22,17 @@ def assert_log_record(alert_group, reason, log_type=AlertGroupLogRecord.TYPE_DIR
         assert log.get_step_specific_info() == expected_info
 
 
-def setup_always_on_call_schedule(make_schedule, make_on_call_shift, organization, team, user, extra_users=None):
-    # setup on call schedule
+@pytest.mark.django_db
+def test_user_is_oncall(make_organization, make_user_for_organization, make_schedule, make_on_call_shift):
+    organization = make_organization()
+    not_oncall_user = make_user_for_organization(organization)
+    oncall_user = make_user_for_organization(organization)
+
+    # set up schedule: user is on call
     schedule = make_schedule(
         organization,
         schedule_class=OnCallScheduleWeb,
-        team=team,
+        team=None,
     )
     now = timezone.now().replace(hour=0, minute=0, second=0, microsecond=0)
     start_date = now - timezone.timedelta(days=7)
@@ -36,91 +47,11 @@ def setup_always_on_call_schedule(make_schedule, make_on_call_shift, organizatio
     on_call_shift = make_on_call_shift(
         organization=organization, shift_type=CustomOnCallShift.TYPE_ROLLING_USERS_EVENT, **data
     )
-    on_call_shift.add_rolling_users([[user]])
-    if extra_users:
-        # add old shifts for users
-        for i, u in enumerate(extra_users):
-            start_date = now - timezone.timedelta(days=14)
-            data = {
-                "start": start_date + timezone.timedelta(hours=i),
-                "rotation_start": start_date,
-                "duration": timezone.timedelta(hours=1),
-                "priority_level": 1,
-                "frequency": CustomOnCallShift.FREQUENCY_DAILY,
-                "schedule": schedule,
-                "until": start_date + timezone.timedelta(days=7),
-            }
-            on_call_shift = make_on_call_shift(
-                organization=organization, shift_type=CustomOnCallShift.TYPE_ROLLING_USERS_EVENT, **data
-            )
-            on_call_shift.add_rolling_users([[u]])
-
+    on_call_shift.add_rolling_users([[oncall_user]])
     schedule.refresh_ical_file()
-    return schedule
 
-
-@pytest.mark.django_db
-def test_check_user_availability_no_policies(make_organization, make_user_for_organization):
-    organization = make_organization()
-    user = make_user_for_organization(organization)
-
-    warnings = check_user_availability(user)
-    assert warnings == [
-        {"data": {}, "error": PagingError.USER_HAS_NO_NOTIFICATION_POLICY},
-        {"data": {"schedules": {}}, "error": PagingError.USER_IS_NOT_ON_CALL},
-    ]
-
-
-@pytest.mark.django_db
-def test_check_user_availability_not_on_call(
-    make_organization, make_user_for_organization, make_user_notification_policy, make_schedule, make_on_call_shift
-):
-    organization = make_organization()
-    user = make_user_for_organization(organization)
-    other_user = make_user_for_organization(organization)
-    make_user_notification_policy(
-        user=user,
-        step=UserNotificationPolicy.Step.NOTIFY,
-        notify_by=UserNotificationPolicy.NotificationChannel.SMS,
-    )
-
-    # setup on call schedule
-    schedule = setup_always_on_call_schedule(
-        make_schedule, make_on_call_shift, organization, None, other_user, extra_users=[user]
-    )
-
-    warnings = check_user_availability(user)
-    assert warnings == [
-        {
-            "data": {"schedules": {schedule.name: {other_user.public_primary_key}}},
-            "error": PagingError.USER_IS_NOT_ON_CALL,
-        },
-    ]
-
-
-@pytest.mark.django_db
-def test_check_user_availability_on_call(
-    make_organization,
-    make_team,
-    make_user_for_organization,
-    make_user_notification_policy,
-    make_schedule,
-    make_on_call_shift,
-):
-    organization = make_organization()
-    some_team = make_team(organization)
-    user = make_user_for_organization(organization)
-    make_user_notification_policy(
-        user=user,
-        step=UserNotificationPolicy.Step.NOTIFY,
-        notify_by=UserNotificationPolicy.NotificationChannel.SMS,
-    )
-
-    # setup on call schedule
-    setup_always_on_call_schedule(make_schedule, make_on_call_shift, organization, some_team, user)
-
-    warnings = check_user_availability(user)
-    assert warnings == []
+    assert user_is_oncall(not_oncall_user) is False
+    assert user_is_oncall(oncall_user) is True
 
 
 @pytest.mark.django_db
@@ -129,65 +60,98 @@ def test_direct_paging_user(make_organization, make_user_for_organization):
     user = make_user_for_organization(organization)
     other_user = make_user_for_organization(organization)
     from_user = make_user_for_organization(organization)
+    msg = "Fire"
 
     with patch("apps.alerts.paging.notify_user_task") as notify_task:
-        direct_paging(
-            organization, None, from_user, title="Help!", message="Fire", users=[(user, False), (other_user, True)]
-        )
+        direct_paging(organization, from_user, msg, users=[(user, False), (other_user, True)])
 
     # alert group created
     alert_groups = AlertGroup.objects.all()
     assert alert_groups.count() == 1
     ag = alert_groups.get()
     alert = ag.alerts.get()
-    assert alert.title == "Help!"
-    assert alert.message == "Fire"
+
+    assert alert.title == f"{from_user.username} is paging {user.username} and {other_user.username} to join escalation"
+    assert alert.message == msg
+
     # notifications sent
     for u, important in ((user, False), (other_user, True)):
         assert notify_task.apply_async.called_with(
             (u.pk, ag.pk), {"important": important, "notify_even_acknowledged": True, "notify_anyway": True}
         )
-        expected_info = {"user": u.public_primary_key, "schedule": None, "important": important}
+        expected_info = {"user": u.public_primary_key, "important": important}
         assert_log_record(ag, f"{from_user.username} paged user {u.username}", expected_info=expected_info)
 
 
 @pytest.mark.django_db
-def test_direct_paging_schedule(
-    make_organization, make_team, make_user_for_organization, make_schedule, make_on_call_shift
-):
+def test_direct_paging_team(make_organization, make_team, make_user_for_organization):
     organization = make_organization()
-    some_team = make_team(organization)
     from_user = make_user_for_organization(organization)
-    user = make_user_for_organization(organization)
-    other_user = make_user_for_organization(organization)
+    team = make_team(organization)
+    msg = "Fire"
 
-    # setup on call schedule
-    schedule = setup_always_on_call_schedule(make_schedule, make_on_call_shift, organization, some_team, user)
-    other_schedule = setup_always_on_call_schedule(
-        make_schedule, make_on_call_shift, organization, some_team, other_user
-    )
-
-    with patch("apps.alerts.paging.notify_user_task") as notify_task:
-        direct_paging(organization, None, from_user, schedules=[(schedule, False), (other_schedule, True)])
+    direct_paging(organization, from_user, msg, team=team)
 
     # alert group created
     alert_groups = AlertGroup.objects.all()
     assert alert_groups.count() == 1
     ag = alert_groups.get()
     alert = ag.alerts.get()
-    assert alert.title == f"Message from {from_user.username}"
-    assert alert.message is None
-    assert_log_record(ag, f"{from_user.username} paged schedule {schedule.name}")
-    assert_log_record(ag, f"{from_user.username} paged schedule {other_schedule.name}")
-    # notifications sent
-    for u, important, s in ((user, False, schedule), (other_user, True, other_schedule)):
-        assert notify_task.apply_async.called_with(
-            (u.pk, ag.pk), {"important": important, "notify_even_acknowledged": True, "notify_anyway": True}
-        )
-        expected_info = {"user": u.public_primary_key, "schedule": s.public_primary_key, "important": important}
-        assert_log_record(
-            ag, f"{from_user.username} paged user {u.username} (from schedule {s.name})", expected_info=expected_info
-        )
+    assert alert.title == f"{from_user.username} is paging {team.name} to join escalation"
+    assert alert.message == msg
+
+    assert ag.channel.verbal_name == f"Direct paging ({team.name} team)"
+    assert ag.channel.team == team
+
+
+@pytest.mark.django_db
+def test_direct_paging_no_team(make_organization, make_user_for_organization):
+    organization = make_organization()
+    from_user = make_user_for_organization(organization)
+    other_user = make_user_for_organization(organization)
+    msg = "Fire"
+
+    direct_paging(organization, from_user, msg, users=[(other_user, False)])
+
+    # alert group created
+    alert_groups = AlertGroup.objects.all()
+    assert alert_groups.count() == 1
+    ag = alert_groups.get()
+    alert = ag.alerts.get()
+    assert alert.title == f"{from_user.username} is paging {other_user.username} to join escalation"
+    assert alert.message == msg
+
+    assert ag.channel.verbal_name == "Direct paging (No team)"
+    assert ag.channel.team is None
+
+
+@pytest.mark.django_db
+def test_direct_paging_custom_title(make_organization, make_user_for_organization):
+    organization = make_organization()
+    from_user = make_user_for_organization(organization)
+    other_user = make_user_for_organization(organization)
+    custom_title = "Custom title"
+    msg = "Fire"
+
+    direct_paging(organization, from_user, msg, custom_title, users=[(other_user, False)])
+
+    # alert group created
+    alert_groups = AlertGroup.objects.all()
+    assert alert_groups.count() == 1
+    ag = alert_groups.get()
+
+    assert ag.web_title_cache == custom_title
+    assert ag.alerts.get().title == custom_title
+
+
+@pytest.mark.django_db
+def test_direct_paging_no_team_and_no_users(make_organization, make_user_for_organization):
+    organization = make_organization()
+    from_user = make_user_for_organization(organization)
+    msg = "Fire"
+
+    with pytest.raises(DirectPagingUserTeamValidationError):
+        direct_paging(organization, from_user, msg)
 
 
 @pytest.mark.django_db
@@ -201,52 +165,18 @@ def test_direct_paging_reusing_alert_group(
     alert_group = make_alert_group(alert_receive_channel=alert_receive_channel)
 
     with patch("apps.alerts.paging.notify_user_task") as notify_task:
-        direct_paging(organization, None, from_user, users=[(user, False)], alert_group=alert_group)
+        direct_paging(organization, from_user, "Fire!", users=[(user, False)], alert_group=alert_group)
 
     # no new alert group is created
     alert_groups = AlertGroup.objects.all()
     assert alert_groups.count() == 1
     assert_log_record(alert_group, f"{from_user.username} paged user {user.username}")
+
     # notifications sent
     ag = alert_groups.get()
     assert notify_task.apply_async.called_with(
         (user.pk, ag.pk), {"important": False, "notify_even_acknowledged": True, "notify_anyway": True}
     )
-
-
-@pytest.mark.django_db
-def test_direct_paging_reusing_alert_group_custom_chain_raises(
-    make_organization, make_user_for_organization, make_alert_receive_channel, make_alert_group, make_escalation_chain
-):
-    organization = make_organization()
-    from_user = make_user_for_organization(organization)
-    alert_receive_channel = make_alert_receive_channel(organization=organization)
-    alert_group = make_alert_group(alert_receive_channel=alert_receive_channel)
-    custom_chain = make_escalation_chain(organization)
-
-    with pytest.raises(ValueError):
-        direct_paging(organization, None, from_user, alert_group=alert_group, escalation_chain=custom_chain)
-
-
-@pytest.mark.django_db
-def test_direct_paging_custom_chain(
-    make_organization, make_user_for_organization, make_alert_receive_channel, make_alert_group, make_escalation_chain
-):
-    organization = make_organization()
-    from_user = make_user_for_organization(organization)
-    custom_chain = make_escalation_chain(organization)
-
-    direct_paging(organization, None, from_user, escalation_chain=custom_chain)
-
-    # alert group created
-    alert_groups = AlertGroup.objects.all()
-    assert alert_groups.count() == 1
-    ag = alert_groups.get()
-    channel_filter = ag.channel_filter_with_respect_to_escalation_snapshot
-    assert channel_filter is not None
-    assert not channel_filter.is_default
-    assert channel_filter.notify_in_slack
-    assert ag.escalation_chain_with_respect_to_escalation_snapshot == custom_chain
 
 
 @pytest.mark.django_db
@@ -256,7 +186,7 @@ def test_direct_paging_returns_alert_group(make_organization, make_user_for_orga
     from_user = make_user_for_organization(organization)
 
     with patch("apps.alerts.paging.notify_user_task"):
-        alert_group = direct_paging(organization, None, from_user, title="Help!", message="Fire", users=[(user, False)])
+        alert_group = direct_paging(organization, from_user, "Help!", users=[(user, False)])
 
     # check alert group returned by direct paging is the same as the one created
     assert alert_group == AlertGroup.objects.get()
@@ -301,15 +231,18 @@ def test_direct_paging_always_create_group(make_organization, make_user_for_orga
     organization = make_organization()
     user = make_user_for_organization(organization)
     from_user = make_user_for_organization(organization)
+    msg = "Help!"
+    users = [(user, False)]
 
     with patch("apps.alerts.paging.notify_user_task") as notify_task:
         # although calling twice with same params, there should be 2 alert groups
-        direct_paging(organization, None, from_user, title="Help!", users=[(user, False)])
-        direct_paging(organization, None, from_user, title="Help!", users=[(user, False)])
+        direct_paging(organization, from_user, msg, users=users)
+        direct_paging(organization, from_user, msg, users=users)
 
     # alert group created
     alert_groups = AlertGroup.objects.all()
     assert alert_groups.count() == 2
+
     # notifications sent
     assert notify_task.apply_async.called_with(
         (user.pk, alert_groups[0].pk), {"important": False, "notify_even_acknowledged": True, "notify_anyway": True}
@@ -317,3 +250,109 @@ def test_direct_paging_always_create_group(make_organization, make_user_for_orga
     assert notify_task.apply_async.called_with(
         (user.pk, alert_groups[1].pk), {"important": False, "notify_even_acknowledged": True, "notify_anyway": True}
     )
+
+
+@pytest.mark.django_db
+def test_construct_title(make_organization, make_team, make_user_for_organization):
+    organization = make_organization()
+    from_user = make_user_for_organization(organization)
+    user1 = make_user_for_organization(organization)
+    user2 = make_user_for_organization(organization)
+    user3 = make_user_for_organization(organization)
+    team = make_team(organization)
+
+    def _title(middle_portion: str) -> str:
+        return f"{from_user.username} is paging {middle_portion} to join escalation"
+
+    one_user = [(user1, False)]
+    two_users = [(user1, False), (user2, True)]
+    multiple_users = two_users + [(user3, False)]
+
+    # no team specified + one user
+    assert _construct_title(from_user, None, one_user) == _title(user1.username)
+
+    # no team specified + two users
+    assert _construct_title(from_user, None, two_users) == _title(f"{user1.username} and {user2.username}")
+
+    # no team specified + multiple users
+    assert _construct_title(from_user, None, multiple_users) == _title(
+        f"{user1.username}, {user2.username} and {user3.username}"
+    )
+
+    # team specified + no users
+    assert _construct_title(from_user, team, []) == _title(team.name)
+
+    # team specified + one user
+    assert _construct_title(from_user, team, one_user) == _title(f"{team.name} and {user1.username}")
+
+    # team specified + two users
+    assert _construct_title(from_user, team, two_users) == _title(f"{team.name}, {user1.username} and {user2.username}")
+
+    # team specified + multiple users
+    assert _construct_title(from_user, team, multiple_users) == _title(
+        f"{team.name}, {user1.username}, {user2.username} and {user3.username}"
+    )
+
+
+@pytest.mark.django_db
+def test_integration_is_notifiable(
+    make_organization,
+    make_alert_receive_channel,
+    make_channel_filter,
+    make_escalation_chain,
+    make_slack_team_identity,
+    make_telegram_channel,
+):
+    organization = make_organization()
+
+    # integration has no default channel filter
+    arc = make_alert_receive_channel(organization)
+    make_channel_filter(arc, is_default=False)
+    assert integration_is_notifiable(arc) is False
+
+    # integration has more than one channel filter
+    arc = make_alert_receive_channel(organization)
+    make_channel_filter(arc, is_default=False)
+    make_channel_filter(arc, is_default=False)
+    assert integration_is_notifiable(arc) is True
+
+    # integration's default channel filter is setup to notify via slack but Slack is not configured for the org
+    arc = make_alert_receive_channel(organization)
+    make_channel_filter(arc, is_default=True, notify_in_slack=True)
+    assert integration_is_notifiable(arc) is False
+
+    # integration's default channel filter is setup to notify via slack and Slack is configured for the org
+    arc = make_alert_receive_channel(organization)
+    slack_team_identity = make_slack_team_identity()
+    organization.slack_team_identity = slack_team_identity
+    organization.save()
+
+    make_channel_filter(arc, is_default=True, notify_in_slack=True)
+    assert integration_is_notifiable(arc) is True
+
+    # integration's default channel filter is setup to notify via telegram but Telegram is not configured for the org
+    arc = make_alert_receive_channel(organization)
+    make_channel_filter(arc, is_default=True, notify_in_slack=False, notify_in_telegram=True)
+    assert integration_is_notifiable(arc) is False
+
+    # integration's default channel filter is setup to notify via telegram and Telegram is configured for the org
+    arc = make_alert_receive_channel(organization)
+    make_channel_filter(arc, is_default=True, notify_in_slack=False, notify_in_telegram=True)
+    make_telegram_channel(organization)
+    assert integration_is_notifiable(arc) is True
+
+    # integration's default channel filter is contactable via a custom messaging backend
+    arc = make_alert_receive_channel(organization)
+    make_channel_filter(
+        arc,
+        is_default=True,
+        notify_in_slack=False,
+        notification_backends={"MSTEAMS": {"channel": "test", "enabled": True}},
+    )
+    assert integration_is_notifiable(arc) is True
+
+    # integration's default channel filter has an escalation chain attached to it
+    arc = make_alert_receive_channel(organization)
+    escalation_chain = make_escalation_chain(organization)
+    make_channel_filter(arc, is_default=True, notify_in_slack=False, escalation_chain=escalation_chain)
+    assert integration_is_notifiable(arc) is True