+++ engine/apps/api/serializers/team.py
@@ -1,10 +1,28 @@
+import typing
+
 from rest_framework import serializers
 
+from apps.schedules.ical_utils import SchedulesOnCallUsers
 from apps.user_management.models import Team
 
 
-class TeamSerializer(serializers.ModelSerializer):
+class TeamSerializerContext(typing.TypedDict):
+    schedules_with_oncall_users: SchedulesOnCallUsers
+
+
+class FastTeamSerializer(serializers.ModelSerializer):
     id = serializers.CharField(read_only=True, source="public_primary_key")
+
+    class Meta:
+        model = Team
+        fields = ["id", "name", "email", "avatar_url"]
+
+
+class TeamSerializer(serializers.ModelSerializer):
+    context: TeamSerializerContext
+
+    id = serializers.CharField(read_only=True, source="public_primary_key")
+    number_of_users_currently_oncall = serializers.SerializerMethodField()
 
     class Meta:
         model = Team
@@ -14,6 +32,7 @@ class TeamSerializer(serializers.ModelSerializer):
             "email",
             "avatar_url",
             "is_sharing_resources_to_all",
+            "number_of_users_currently_oncall",
         )
 
         read_only_fields = [
@@ -22,3 +41,12 @@ class TeamSerializer(serializers.ModelSerializer):
             "email",
             "avatar_url",
         ]
+
+    def get_number_of_users_currently_oncall(self, obj: Team) -> int:
+        num_of_users_oncall_for_team = 0
+
+        for schedule, users in self.context["schedules_with_oncall_users"].items():
+            if schedule.team == obj:
+                num_of_users_oncall_for_team += len(users)
+
+        return num_of_users_oncall_for_team