+++ grafana-plugin/src/pages/incidents/Incidents.tsx
@@ -118,7 +118,7 @@ class Incidents extends React.Component<IncidentsPageProps, IncidentsPageState> 
               <Text.Title level={3}>Alert Groups</Text.Title>
               <WithPermissionControlTooltip userAction={UserActions.AlertGroupsDirectPaging}>
                 <Button icon="plus" onClick={this.handleOnClickEscalateTo}>
-                  New alert group
+                  Escalation
                 </Button>
               </WithPermissionControlTooltip>
             </HorizontalGroup>